Separate downloader from the manager. Optimize folder lookups
