+++ app/src/main/java/eu/kanade/tachiyomi/ui/recent_updates/RecentChaptersPresenter.kt
@@ -1,6 +1,7 @@
 package eu.kanade.tachiyomi.ui.recent_updates
 
 import android.os.Bundle
+import com.hippo.unifile.UniFile
 import eu.kanade.tachiyomi.data.database.DatabaseHelper
 import eu.kanade.tachiyomi.data.database.models.MangaChapter
 import eu.kanade.tachiyomi.data.download.DownloadManager
@@ -97,7 +98,10 @@ class RecentChaptersPresenter : BasePresenter<RecentChaptersFragment>() {
                 .map { mangaChapters ->
                     mangaChapters.map { it.toModel() }
                 }
-                .doOnNext { chapters = it }
+                .doOnNext {
+                    setDownloadedChapters(it)
+                    chapters = it
+                }
                 // Group chapters by the date they were fetched on a ordered map.
                 .flatMap { recentItems ->
                     Observable.from(recentItems)
@@ -142,16 +146,27 @@ class RecentChaptersPresenter : BasePresenter<RecentChaptersFragment>() {
         // downloaded and assign it to the status.
         if (download != null) {
             model.download = download
-        } else {
-            // Get source of chapter.
-            val source = sourceManager.get(manga.source)!!
-
-            model.status = if (downloadManager.isChapterDownloaded(source, manga, chapter))
-                Download.DOWNLOADED
-            else
-                Download.NOT_DOWNLOADED
         }
         return model
+    }
+
+    /**
+     * Finds and assigns the list of downloaded chapters.
+     *
+     * @param chapters the list of chapter from the database.
+     */
+    private fun setDownloadedChapters(chapters: List<RecentChapter>) {
+        val cachedDirs = mutableMapOf<Long, UniFile?>()
+
+        chapters.forEach { chapter ->
+            val manga = chapter.manga
+            val mangaDir = cachedDirs.getOrPut(manga.id!!)
+                    { downloadManager.findMangaDir(sourceManager.get(manga.source)!!, manga) }
+
+            if (mangaDir?.findFile(downloadManager.getChapterDirName(chapter)) != null) {
+                chapter.status = Download.DOWNLOADED
+            }
+        }
     }
 
     /**
